Ensure `nullable` works for allOf, anyOf etc.
Previously the `nullable` check was only being applied to the `type`
attribute validation, but schemas using `anyOf`, `allOf`, `oneOf` etc
were unable to use `nullable` because they would never be passed
directly to the field validator. This is crucial when using `$ref`
combined with `nullable`, because these cases require the use of
`allOf`.

The solution here is to simplify the validation extension at the root
and not to attach the `nullable` handling to any particular attribute.

Before this change, the following would fail validation if `foo` was
`null`:

```
{
  "type": "object",
  "properties": {
    "foo": {
      "nullable": true,
      "allOf": [
        {
          "$ref": "#/components/schemas/Foo"
        }
      ]
    }
  }
}
```

With this change the `nullable` check is handled at the right spot
before the `allOf` is evaluated.